Use Float instead of Double for all the level metering logic.
This is in preparation of further changes to the way the audio level is computed and to allow recording and sending of the waveform.
The main reasoning behind the change is twofold:
1) We don't need the precision of Double in our context (we just need a rough indication of the changes in audio level to successfully draw a level meter or a waveform in our UI).
2) Performance: It is true that on 64 bit CPUs single operations involving Floats or Doubles take the same amount of time (i.e one clock cycle). But there are other aspects here that vouch in favor of Floats:
	- A float takes half the space in memory compared to a double, so when storing long lists of them this can add up.
	- On Android O and greater the ART runtime can "vectorize" certain operations on lists and make use of the CPUs SIMD registers which are generally 128 bits. So by using floats 4 of them can fit and be computed at the same time whilst with doubles only 2 will fit halving the throughput.

References:
- https://source.android.com/docs/core/runtime/improvements
- https://www.slideshare.net/linaroorg/automatic-vectorization-in-art-android-runtime-sfo17216

+++ libraries/voicerecorder/impl/src/test/kotlin/io/element/android/libraries/voicerecorder/test/FakeAudioLevelCalculator.kt
@@ -19,8 +19,8 @@ package io.element.android.libraries.voicerecorder.test
 import io.element.android.libraries.voicerecorder.impl.audio.AudioLevelCalculator
 import kotlin.math.abs
 
-class FakeAudioLevelCalculator: AudioLevelCalculator {
-    override fun calculateAudioLevel(buffer: ShortArray): Double {
-        return buffer.map { abs(it.toDouble()) }.average() / Short.MAX_VALUE
+class FakeAudioLevelCalculator : AudioLevelCalculator {
+    override fun calculateAudioLevel(buffer: ShortArray): Float {
+        return buffer.map { abs(it.toFloat()) }.average().toFloat() / Short.MAX_VALUE
     }
 }
